Fixup types

+++ rs/ic_os/build_tools/partition_tools/src/ext.rs
@@ -25,7 +25,7 @@ pub struct ExtPartition {
 #[async_trait]
 impl Partition for ExtPartition {
     /// Open an ext4 partition for writing, via debugfs
-    async fn open(image: PathBuf, index: Option<usize>) -> Result<Self> {
+    async fn open(image: PathBuf, index: Option<u32>) -> Result<Self> {
         let _ = debugfs().context("debugfs is needed to open ext4 partitions")?;
 
         if let Some(index) = index {
